Fix functions after the move
